refactor: archiveVIew 에서 GridView 관련 리팩토링
